Removed some duplicate code and restructured it
