+++ addon/pycThermopack/pyctp/map_platform_specifics.py
@@ -5,6 +5,7 @@ from os import path
 import itertools
 # Import sysconfig to detect mingw environment
 import sysconfig
+from datetime import datetime
 
 # GNU FORTRAN
 G_PREFIX = "__"
@@ -21,7 +22,7 @@ I_POSTFIX_NM = "_"
 def get_platform_specifics_from_platform():
     """Get platform specific stuff."""
 
-    #Setting GNU FORTRAN as default
+    # Setting GNU FORTRAN as default
     platform_specifics = {}
     platform_specifics["os_id"] = ""
     platform_specifics["prefix"] = G_PREFIX
@@ -123,3 +124,26 @@ def get_platform_specifics_by_trial_and_error():
         platform_specifics["os_id"] = "win"
 
     return platform_specifics
+
+
+def write_platform_specifics_file(pf_specifics, filename):
+    """Write file for platform specifics"""
+    lines = []
+    lines.append(
+        "# Module for platform specific stuff. Automatically generated.")
+    lines.append("# Timestamp : " +
+                 str(datetime.today().isoformat()) + "\n\n")
+
+    tab = " "*4
+    lines.append("def get_platform_specifics():")
+    lines.append(tab + "pf_specifics = {}")
+
+    for k, v in pf_specifics.items():
+        lines.append(tab + 'pf_specifics["'+k+'"] = "'+v+'"')
+
+    lines.append(tab + "return pf_specifics")
+
+    with open(filename, "w") as f:
+        for line in lines:
+            f.write(line)
+            f.write("\n")
